Fix some NMI timing tests
CPU _should_ check for interrupts at the start of the last cycle of an instruction, because we check at the end we may catch an interrupt triggered by the last cycle of the CPU instruction. This resolves that timing issue by tracking the cycle count (in ppu cycles) at the time of the interrupt firing.

Only implemented for NMI, IRQ will be later