Make the way that interrupts are polled more accurate
Specifically this aims to resolve the issue where interrupts need to be polled at the end of the previous cycle rather than the start of the next.

There's only one test that is added (CLI latency) and all other interrupt tests are currently failing. Very hard to say why at this point but it seems to be something subtle with when NMIs are scheduled around vblank and reading from PPUSTATUS. I swear it's correct though!

Fixes #37